Refactor manifest structure
This changes the way manifests are kept track of.
The default manifest now points to the latest released version of seL4
and libraries.  This commit will be tagged with the current version.
As new versions are released, instead of creating new manifest files,
the default.xml will be updated with the new versions and a new tag will
be created.  In order to refer to older versions, repo can be told an
older tag by adding `-b refs/tags/6.0.0` to the repo init command.  To
refer to nightly tips, master.xml should now be used.  The same
applies for mcs and cmake variants

There may be more information at https://wiki.sel4.systems/ReleaseProcess